requirements.txt
beautifulsoup4==4.6.3
requests==2.21.0
openpyxl==3.0.1
pandas==0.24.2
PyYAML==5.3
